website/requirements.txt
amqp==1.4.9
anyjson==0.3.3
six==1.10.0
beautifulsoup4==4.3.2
billiard==3.3.0.23
celery==3.1.18
cffi==1.9.1
cryptography==1.7
dj-database-url==0.3.0
Django==1.11.4
django-allauth==0.29.0
django-appconf==1.0.1
django-celery==3.1.16
django-celery-email==1.1.5
django-contact-form==1.0
django-cors-headers==1.3.1
django-countries==3.4.1
django-dajax==0.9.2
django-dajaxice==0.7
django-filter==0.14.0
django-form-utils==1.0.1
django-formtools==1.0
django-genericadmin==0.6.1
django-label-tag-attr==1.0.2
django-localflavor==1.3
django-markup==0.4
django-mptt==0.8.7
django-mptt-admin==0.3.1
django-otp==0.3.3
django-recaptcha==1.0.4
django-rest-auth==0.8.2
django-staticfiles==1.2.1
django-two-factor-auth==1.2.2
django-widget-tweaks==1.4.1
djangorestframework==3.4.6
djangorestframework-recursive==0.1.1
enum34==1.1.6
facebook-sdk==1.0.0
flake8==2.4.1
future==0.16.0
gunicorn==19.1.1
html2text==2014.12.5
httplib2==0.9.2
hurry.filesize==0.9
idna==2.1
ipaddress==1.0.17
kombu==3.0.37
Markdown==2.6.5
mccabe==0.3.1
mistune==0.7.1
MySQL-python==1.2.5
oauth2==1.9.0.post1
oauth2client==1.5.2
oauthlib==1.0.3
paypalrestsdk==1.12.0
pep8==1.5.7
Pillow==2.6.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycurl==7.43.0
pyflakes==0.8.1
Pygments==2.0.1
pyOpenSSL==16.2.0
python-dateutil==2.4.2
python-instagram==1.3.2
python-openid==2.2.5
python-twitter==3.2
pytz==2017.2
qrcode==4.0.4
redis==2.10.5
request==0.0.2
requests==2.8.1
requests-oauthlib==0.6.1
rsa==3.4.2
simplejson==3.8.0
urllib3==1.12
vine==1.1.3
